Move static Log classes out of generic types
