+++ src/Grpc.Net.Client/Internal/HttpContentClientStreamReader.cs
@@ -25,6 +25,7 @@ using System.Threading.Tasks;
 using Grpc.Core;
 using Grpc.Shared;
 using Microsoft.Extensions.Logging;
+using Log = Grpc.Net.Client.Internal.HttpContentClientStreamReaderLog;
 
 namespace Grpc.Net.Client.Internal
 {
@@ -34,8 +35,6 @@ namespace Grpc.Net.Client.Internal
     {
         // Getting logger name from generic type is slow. Cached copy.
         private const string LoggerName = "Grpc.Net.Client.Internal.HttpContentClientStreamReader";
-
-        private static readonly Task<bool> FinishedTask = Task.FromResult(false);
 
         private readonly GrpcCall<TRequest, TResponse> _call;
         private readonly ILogger _logger;
@@ -90,7 +89,7 @@ namespace Grpc.Net.Client.Internal
                 if (status.StatusCode == StatusCode.OK)
                 {
                     // Response is finished and it was successful so just return false
-                    return FinishedTask;
+                    return CommonGrpcProtocolHelpers.FalseTask;
                 }
                 else
                 {
@@ -247,16 +246,16 @@ namespace Grpc.Net.Client.Internal
                 return moveNextTask != null && !moveNextTask.IsCompleted;
             }
         }
+    }
 
-        private static class Log
+    internal static class HttpContentClientStreamReaderLog
+    {
+        private static readonly Action<ILogger, Exception> _readMessageError =
+            LoggerMessage.Define(LogLevel.Error, new EventId(1, "ReadMessageError"), "Error reading message.");
+
+        public static void ReadMessageError(ILogger logger, Exception ex)
         {
-            private static readonly Action<ILogger, Exception> _readMessageError =
-                LoggerMessage.Define(LogLevel.Error, new EventId(1, "ReadMessageError"), "Error reading message.");
-
-            public static void ReadMessageError(ILogger logger, Exception ex)
-            {
-                _readMessageError(logger, ex);
-            }
+            _readMessageError(logger, ex);
         }
     }
 }
