+++ src/Grpc.Net.Client/Internal/Retry/HedgingCall.cs
@@ -24,6 +24,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Grpc.Core;
 using Grpc.Shared;
+using Log = Grpc.Net.Client.Internal.Retry.RetryCallBaseLog;
 
 namespace Grpc.Net.Client.Internal.Retry
 {
