+++ src/Grpc.Net.Client/Internal/Retry/RetryCallBaseClientStreamWriter.cs
@@ -21,6 +21,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 using Grpc.Core;
+using Log = Grpc.Net.Client.Internal.ClientStreamWriterBaseLog;
 
 namespace Grpc.Net.Client.Internal.Retry
 {
